ApisAgro-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
google-generativeai
dotenv